Refactor out IO and command parsing. This is to allow alternative UI implementations.
