+++ src/test/java/org/duke/ui/CommandTest.java
@@ -1,5 +1,6 @@
 package org.duke.ui;
 
+import org.duke.cmd.Command;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
